Phase 3 complete: Full price pipeline + arbitrage + set completion APIs
- Fix price ingestion: concurrent inserts + resume support (316K prices ingested)
- Add market snapshot aggregation (134K daily snapshots)
- Add POST /api/v1/arbitrage — US/EU price spread detection
- Add POST /api/v1/set/complete — missing cards + cost estimator
- Fix ROUND() cast for Postgres PERCENTILE_CONT

+++ src/lib/ingestion/aggregate-snapshots.ts
@@ -37,10 +37,10 @@ async function main() {
       pp.collectible_id,
       pp.condition,
       CURRENT_DATE,
-      ROUND(AVG(pp.price_usd::numeric), 2),
-      ROUND(PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY pp.price_usd::numeric), 2),
-      ROUND(MIN(pp.price_usd::numeric), 2),
-      ROUND(MAX(pp.price_usd::numeric), 2),
+      ROUND(AVG(pp.price_usd::numeric)::numeric, 2),
+      ROUND((PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY pp.price_usd::numeric))::numeric, 2),
+      ROUND(MIN(pp.price_usd::numeric)::numeric, 2),
+      ROUND(MAX(pp.price_usd::numeric)::numeric, 2),
       COUNT(*)::int,
       COUNT(DISTINCT pp.source)::int
     FROM price_points pp
